requirements.txt
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-appconf==1.0.3
django-heroku==0.3.1
django-solo==1.1.3
gunicorn==20.0.4
psycopg2==2.8.4
zeep==3.4.0
django-import-export==2.0.2
django-notifications-hq==1.6.0
djangorestframework~=3.11.0